Move read_sysfs_prop() from names.c to its own file
This function was added to names.c by commit dddb696e0172 ("lsusb: Read
unkown names from sysfs device desc."), but it doesn't really belong
there: sysfs is an entirely different data source from the udev hwdb
that names.c works with, and this function both does things (like
directly accessing files) that other functions in names.c don't and
doesn't do things (like depending on the work done by names_init()) that
those other functions do.
